requirements.txt
python-dotenv
pdfplumber